pyproject.toml
[project]
name = "artifact-infer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "flashinfer-python",
    "math-verify>=0.9.0",
    "matplotlib>=3.10.8",
    "psutil>=7.2.1",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "sgl-kernel==0.3.17",
    "torch==2.8",
    "transformers==4.57.1",
    "xxhash>=3.6.0",
    "zmq>=0.0.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.sources]
flashinfer-python = { path = "third_party/flashinfer", editable = true }
